chore: adding Configuration MSBuild project
