+++ src/Mobile.BuildTools.Configuration.MSBuild/ConfigurationManager.targets
