--- src/Mobile.BuildTools/Configuration/MetaData.cs
@@ -1,8 +1,0 @@
-﻿namespace Mobile.BuildTools.Configuration
-{
-    internal static class MetaData
-    {
-        public const string Link = nameof(Link);
-        public const string SourceFile = nameof(SourceFile);
-    }
-}
